requirements.txt
streamlit>=1.28.0
groq>=0.4.0
google-generativeai>=0.5.0
anthropic>=0.18.0
python-dotenv>=1.0.0
pdfplumber>=0.10.0
python-docx>=1.1.0
